backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic==2.5.2
sqlalchemy==2.0.23
python-jose==3.3.0
cors==1.0.1
